refactor: rename price-alert -> price-notification
