+++ prisma/schema.prisma
@@ -14,20 +14,6 @@ datasource db {
 }
 
 /// This model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments
-model AlertHistories {
-  idx         String      @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  userIdx     Int         @map("user_idx")
-  alertIdx    String      @map("alert_idx") @db.Uuid
-  rate        Decimal     @db.Decimal(20, 8)
-  triggeredAt DateTime    @default(now()) @map("triggered_at") @db.Timestamptz(6)
-  priceAlerts PriceAlerts @relation(fields: [alertIdx], references: [idx], onDelete: Cascade, onUpdate: NoAction, map: "fk_price_alerts_to_alert_histories")
-  users       Users       @relation(fields: [userIdx], references: [idx], onDelete: Cascade, onUpdate: NoAction, map: "fk_users_to_alert_histories")
-
-  @@index([alertIdx], map: "idx_alert_histories_alert")
-  @@map("alert_histories")
-}
-
-/// This model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments
 model ExchangeRates {
   idx          Int      @id @default(autoincrement())
   baseCurrency String   @map("base_currency") @db.Char(3)
@@ -59,39 +45,19 @@ model News {
 }
 
 /// This model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments
-model PriceAlerts {
-  idx            String           @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
-  userIdx        Int              @map("user_idx")
-  baseCurrency   String           @map("base_currency") @db.Char(3)
-  currencyCode   String           @map("currency_code") @db.Char(3)
-  targetPrice    Decimal          @map("target_price") @db.Decimal(20, 8)
-  condition      String           @db.VarChar
-  isTriggered    Boolean          @default(false) @map("is_triggered")
-  isRepeatable   Boolean          @default(false) @map("is_repeatable")
-  createdAt      DateTime         @default(now()) @map("created_at") @db.Timestamptz(6)
-  updatedAt      DateTime         @default(now()) @map("updated_at") @db.Timestamptz(6)
-  deletedAt      DateTime?        @map("deleted_at") @db.Timestamptz(6)
-  alertHistories AlertHistories[]
-  users          Users            @relation(fields: [userIdx], references: [idx], onDelete: Cascade, onUpdate: NoAction, map: "fk_users_to_price_alerts")
-
-  @@index([userIdx, currencyCode, baseCurrency], map: "idx_price_alerts_user")
-  @@map("price_alerts")
-}
-
-/// This model or at least one of its fields has comments in the database, and requires an additional setup for migrations: Read more: https://pris.ly/d/database-comments
 model Users {
-  idx            Int              @id @default(autoincrement())
-  email          String           @db.VarChar(255)
-  name           String           @db.VarChar
-  password       String?          @db.Char(60)
-  socialProvider String           @map("social_provider") @db.VarChar(20)
-  socialId       String           @map("social_id") @db.VarChar(100)
-  createdAt      DateTime         @default(now()) @map("created_at") @db.Timestamptz(6)
-  updatedAt      DateTime         @default(now()) @map("updated_at") @db.Timestamptz(6)
-  deletedAt      DateTime?        @map("deleted_at") @db.Timestamptz(6)
-  alertHistories AlertHistories[]
-  priceAlerts    PriceAlerts[]
-  watchlist      Watchlist[]
+  idx                     Int                       @id @default(autoincrement())
+  email                   String                    @db.VarChar(255)
+  name                    String                    @db.VarChar
+  password                String?                   @db.Char(60)
+  socialProvider          String                    @map("social_provider") @db.VarChar(20)
+  socialId                String                    @map("social_id") @db.VarChar(100)
+  createdAt               DateTime                  @default(now()) @map("created_at") @db.Timestamptz(6)
+  updatedAt               DateTime                  @default(now()) @map("updated_at") @db.Timestamptz(6)
+  deletedAt               DateTime?                 @map("deleted_at") @db.Timestamptz(6)
+  notifications_histories notifications_histories[]
+  price_notifications     price_notifications[]
+  watchlist               Watchlist[]
 
   @@unique([email, deletedAt], map: "users_email_unique")
   @@unique([socialProvider, socialId, deletedAt], map: "users_social_unique")
@@ -130,4 +96,34 @@ model ExchangeRatesDaily {
   @@unique([currencyCode, baseCurrency, ohlcDate], map: "exchange_rates_daily_unique")
   @@index([currencyCode, baseCurrency, ohlcDate], map: "idx_exchange_rates_lookup")
   @@map("exchange_rates_daily")
+}
+
+model notifications_histories {
+  idx                 String              @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
+  user_idx            Int
+  notifications_idx   String              @db.Uuid
+  rate                Decimal             @db.Decimal(20, 8)
+  triggered_at        DateTime            @default(now()) @db.Timestamptz(6)
+  price_notifications price_notifications @relation(fields: [notifications_idx], references: [idx], onDelete: Cascade, onUpdate: NoAction, map: "fk_price_notifications_to_notifications_histories")
+  users               Users               @relation(fields: [user_idx], references: [idx], onDelete: Cascade, onUpdate: NoAction, map: "fk_users_to_notifications_histories")
+
+  @@index([notifications_idx], map: "idx_notifications_histories_notifications")
+}
+
+model price_notifications {
+  idx                     String                    @id @default(dbgenerated("gen_random_uuid()")) @db.Uuid
+  user_idx                Int
+  base_currency           String                    @db.Char(3)
+  currency_code           String                    @db.Char(3)
+  target_price            Decimal                   @db.Decimal(20, 8)
+  condition               String                    @db.VarChar
+  is_triggered            Boolean                   @default(false)
+  is_repeatable           Boolean                   @default(false)
+  created_at              DateTime                  @default(now()) @db.Timestamptz(6)
+  updated_at              DateTime                  @default(now()) @db.Timestamptz(6)
+  deleted_at              DateTime?                 @db.Timestamptz(6)
+  notifications_histories notifications_histories[]
+  users                   Users                     @relation(fields: [user_idx], references: [idx], onDelete: Cascade, onUpdate: NoAction, map: "fk_users_to_price_notifications")
+
+  @@index([user_idx, currency_code, base_currency], map: "idx_price_notifications_user")
 }
