+++ src/apis/price-alert/interfaces/price-notification.interface.ts
@@ -1,0 +1,14 @@
+import { PriceNotificationEntity } from '../entitites/price-alert.entity';
+
+export namespace IPriceAlertEntity {
+  export interface ICreate
+    extends Pick<
+      PriceNotificationEntity,
+      | 'userIdx'
+      | 'baseCurrency'
+      | 'currencyCode'
+      | 'targetPrice'
+      | 'condition'
+      | 'isRepeatable'
+    > {}
+}
