--- src/apis/price-alert/price-alert.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PriceAlertService } from './price-alert.service';
-import { PriceAlertController } from './price-alert.controller';
-
-@Module({
-  controllers: [PriceAlertController],
-  providers: [PriceAlertService],
-})
-export class PriceAlertModule {}
