+++ src/apis/price-alert/price-notification.controller.ts
@@ -1,0 +1,11 @@
+import { Controller } from '@nestjs/common';
+import { PriceNotificationService } from './price-notification.service';
+
+@Controller('price-notification')
+export class PriceNotificationController {
+  constructor(private readonly priceAlertService: PriceNotificationService) {}
+
+  /**
+   * Create Notification
+   */
+}
