+++ src/apis/price-alert/price-notification.service.ts
@@ -1,4 +1,4 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class PriceAlertService {}
+export class PriceNotificationService {}
