+++ src/apis/price-alert/repository/price-notification.repository.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@nestjs/common';
+import { PrismaService } from '../../../prisma/prisma.service';
+
+@Injectable()
+export class PriceNotificationRepository {
+  constructor(prisma: PrismaService) {}
+}
